book_database/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.2
pysyncobj==0.3.12